Add args for button and breadcrumb click events
